requirements.txt
flet==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
python-dotenv==1.0.0
requests==2.25.1
